refactor(contacts): move contacts page ui to widget
